Import Insight: * itcl from GDB CVS HEAD ([Incr Tcl], a dependency - also includes iwidgets) * libgui from CVS HEAD (TCL/Tk gui stuff by Red Hat, also a dependency) * gdb/gdbtk (Insight itself) from GDB 6.3 branchpoint, with MinGW patches applied to the library subdirectory (TCL files) and with the plugins removed. I'll take care of porting the C files.
For TCL and Tk, the plan is to rely on a systemwide install. (For Windows, I'll use the binaries from mingw.org.)

WARNING: The branch is not necessarily compilable right now, and there might be compile failures for the next few days, until the Insight integration is completed. I'll do my best to keep the non-compilable revisions to a minimum though.

git-svn-id: http://svn.tilp.info/repos/tiemu@1267 4f589446-7ac1-0310-82c7-a055b08196ef